Introduce a default_stepsize(::AbstractManifold, ::Type{<:AbstractSolverState})
...which is actually very nice since we can have a very unified way of a stepsize even when using the state for a subsolver
Change the Gradient descent default slightly, and adapt the CostIncrease warning slightly.
