+++ src/plans/solver_state.jl
@@ -191,9 +191,10 @@ end
 _set_iterate!(s::AbstractManoptSolverState, M, p, ::Val{true}) = set_iterate!(s.state, M, p)
 
 """
-    get_solver_result(O::AbstractManoptSolverState)
+    get_solver_result(ams::AbstractManoptSolverState)
 
-return the (last stored) iterate within [`AbstractManoptSolverState`](@ref)``O`. By default also undecorates the state beforehand
+Return the final result after all iterations that is stored within
+the [`AbstractManoptSolverState`](@ref) `ams`, which was modified during the iterations.
 """
 function get_solver_result(s::AbstractManoptSolverState)
     return get_solver_result(s, dispatch_state_decorator(s))
