+++ src/solvers/ChambollePock.jl
@@ -1,4 +1,143 @@
 @doc raw"""
+    ChambollePockState <: AbstractPrimalDualSolverState
+
+stores all options and variables within a linearized or exact Chambolle Pock.
+The following list provides the order for the constructor, where the previous iterates are
+initialized automatically and values with a default may be left out.
+
+* `m` - base point on ``\mathcal M``
+* `n` - base point on ``\mathcal N``
+* `p` - an initial point on ``x^{(0)} ∈\mathcal M`` (and its previous iterate)
+* `X` - an initial tangent vector ``X^{(0)}∈T^*\mathcal N`` (and its previous iterate)
+* `pbar` - the relaxed iterate used in the next dual update step (when using `:primal` relaxation)
+* `Xbar` - the relaxed iterate used in the next primal update step (when using `:dual` relaxation)
+* `Θ` – factor to damp the helping ``\tilde x``
+* `primal_stepsize` – (`1/sqrt(8)`) proximal parameter of the primal prox
+* `dual_stepsize` – (`1/sqrt(8)`) proximal parameter of the dual prox
+* `acceleration` – (`0.`) acceleration factor due to Chambolle & Pock
+* `relaxation` – (`1.`) relaxation in the primal relaxation step (to compute `pbar`)
+* `relax` – (`_primal`) which variable to relax (`:primal` or `:dual`)
+* `stop` - a [`StoppingCriterion`](@ref)
+* `type` – (`exact`) whether to perform an `:exact` or `:linearized` Chambolle-Pock
+* `update_primal_base` (`(p,o,i) -> o.m`) function to update the primal base
+* `update_dual_base` (`(p,o,i) -> o.n`) function to update the dual base
+* `retraction_method` – (`default_retraction_method(M)`) the retraction to use
+* `inverse_retraction_method` - (`default_inverse_retraction_method(M)`) an inverse
+  retraction to use on the manifold ``\mathcal M``.
+* `inverse_retraction_method_dual` - (`default_inverse_retraction_method(N)`)
+  an inverse retraction to use on manifold ``\mathcal N``.
+* `vector_transport_method` - (`default_vector_transport_method(M)`) a vector transport to
+  use on the manifold ``\mathcal M``.
+* `vector_transport_method_dual` - (`default_vector_transport_method(N)`) a
+  vector transport to use on manifold ``\mathcal N``.
+
+where for the last two the functions a [`Problem`](@ref)` p`,
+[`AbstractManoptSolverState`](@ref)` o` and the current iterate `i` are the arguments.
+If you activate these to be different from the default identity, you have to provide
+`p.Λ` for the algorithm to work (which might be `missing` in the linearized case).
+
+# Constructor
+    ChambollePockState(M::AbstractManifold,
+        m::P, n::Q, p::P, X::T, primal_stepsize::Float64, dual_stepsize::Float64;
+        kwargs...
+    )
+where all other fields from above are keyword arguments with their default values given in brackets,
+as well as `N=TangentBundle(M)`
+"""
+mutable struct ChambollePockState{
+    P,
+    Q,
+    T,
+    RM<:AbstractRetractionMethod,
+    IRM<:AbstractInverseRetractionMethod,
+    IRM_Dual<:AbstractInverseRetractionMethod,
+    VTM<:AbstractVectorTransportMethod,
+    VTM_Dual<:AbstractVectorTransportMethod,
+} <: AbstractPrimalDualSolverState
+    m::P
+    n::Q
+    p::P
+    pbar::P
+    X::T
+    Xbar::T
+    primal_stepsize::Float64
+    dual_stepsize::Float64
+    acceleration::Float64
+    relaxation::Float64
+    relax::Symbol
+    stop::StoppingCriterion
+    variant::Symbol
+    update_primal_base::Union{Function,Missing}
+    update_dual_base::Union{Function,Missing}
+    retraction_method::RM
+    inverse_retraction_method::IRM
+    inverse_retraction_method_dual::IRM_Dual
+    vector_transport_method::VTM
+    vector_transport_method_dual::VTM_Dual
+
+    function ChambollePockState(
+        M::AbstractManifold,
+        m::P,
+        n::Q,
+        p::P,
+        X::T;
+        N=TangentBundle(M),
+        primal_stepsize::Float64=1 / sqrt(8),
+        dual_stepsize::Float64=1 / sqrt(8),
+        acceleration::Float64=0.0,
+        relaxation::Float64=1.0,
+        relax::Symbol=:primal,
+        stopping_criterion::StoppingCriterion=StopAfterIteration(300),
+        variant::Symbol=:exact,
+        update_primal_base::Union{Function,Missing}=missing,
+        update_dual_base::Union{Function,Missing}=missing,
+        retraction_method::RM=default_retraction_method(M),
+        inverse_retraction_method::IRM=default_inverse_retraction_method(M),
+        inverse_retraction_method_dual::IRM_Dual=default_inverse_retraction_method(N),
+        vector_transport_method::VTM=default_vector_transport_method(M),
+        vector_transport_method_dual::VTM_Dual=default_vector_transport_method(N),
+    ) where {
+        P,
+        Q,
+        T,
+        RM<:AbstractRetractionMethod,
+        IRM<:AbstractInverseRetractionMethod,
+        IRM_Dual<:AbstractInverseRetractionMethod,
+        VTM<:AbstractVectorTransportMethod,
+        VTM_Dual<:AbstractVectorTransportMethod,
+    }
+        return new{P,Q,T,RM,IRM,IRM_Dual,VTM,VTM_Dual}(
+            m,
+            n,
+            p,
+            copy(M, p),
+            X,
+            copy(N, X),
+            primal_stepsize,
+            dual_stepsize,
+            acceleration,
+            relaxation,
+            relax,
+            stopping_criterion,
+            variant,
+            update_primal_base,
+            update_dual_base,
+            retraction_method,
+            inverse_retraction_method,
+            inverse_retraction_method_dual,
+            vector_transport_method,
+            vector_transport_method_dual,
+        )
+    end
+end
+get_solver_result(apds::AbstractPrimalDualSolverState) = get_iterate(apds)
+get_iterate(apds::AbstractPrimalDualSolverState) = apds.p
+function set_iterate!(apds::AbstractPrimalDualSolverState, p)
+    apds.p = p
+    return apds
+end
+
+@doc raw"""
     ChambollePock(
         M, N, cost, x0, ξ0, m, n, prox_F, prox_G_dual, adjoint_linear_operator;
         forward_operator=missing,
