+++ src/solvers/augmented_Lagrangian_method.jl
@@ -249,7 +249,7 @@ function augmented_Lagrangian_method!(
                 M, copy(M, p), InverseBFGS(), 30
             ),
             stopping_criterion=sub_stopping_criterion,
-            stepsize=WolfePowellLinesearch(M, 1e-3, 0.999; linesearch_stopsize=1e-8),
+            stepsize=default_stepsize(M, QuasiNewtonState),
         );
         sub_kwargs...,
     ),
