+++ test/plans/test_stopping_criteria.jl
@@ -100,7 +100,7 @@ end
 @testset "Stop with step size" begin
     mgo = ManifoldGradientObjective((M, x) -> x^2, x -> 2x)
     dmp = DefaultManoptProblem(Euclidean(), mgo)
-    gds = GradientDescentState(Euclidean(), 1.0)
+    gds = GradientDescentState(Euclidean(), 1.0; stepsize=ConstantStepsize(Euclidean()))
     s1 = StopWhenStepsizeLess(0.5)
     @test !s1(dmp, gds, 1)
     @test s1.reason == ""
